Adds three APIs, and implements them on X11, Cocoa, and Windows:

- SDL_CaptureMouse()
- SDL_GetGlobalMouseState()
- SDL_SetWindowHitTest()

+++ src/video/cocoa/SDL_cocoamessagebox.m
@@ -132,7 +132,7 @@ Cocoa_ShowMessageBox(const SDL_MessageBoxData *messageboxdata, int *buttonid)
     }
     else
     {
-        returnValue = SDL_SetError("Did not get a valid `clicked button' id: %d", clicked);
+        returnValue = SDL_SetError("Did not get a valid `clicked button' id: %ld", (long)clicked);
     }
 
     [pool release];
